+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Conventions / labels:
 - support multiple timeoutTrigs instances in http service so timeouts do not conflict with or clear each other if there are multiple http calls running at once [core-default, core-scss]
 - add layout directive (move resize / DOM stuff from LayoutCtrl to here instead) [core-default, core-scss]
 - package.json - move grunt, etc. to devDependencies [core-default, core-scss]
+- add backend realtime service with socket.io for out of the box web socket / realtime (still need to enable on frontend in index.html) [core-default, core-scss]
 
 
 # 1.0.7 (2013-12-02)